[Refactor] Clean up router types
