+++ src/Components/__tests__/Artsy2.test.tsx
@@ -42,7 +42,6 @@ describe("Artsy context", () => {
         <Artsy.ContextConsumer>
           {props => {
             expect(Object.keys(props).sort()).toEqual([
-              "_isNestedInProvider",
               "currentUser",
               "relayEnvironment",
             ])
