+++ src/Router/Components/Boot.tsx
@@ -3,7 +3,6 @@ import { track } from "Analytics"
 import React from "react"
 import { Environment } from "relay-runtime"
 import * as Artsy from "Router/Artsy2"
-import { MatchingMediaQueries } from "Router/types"
 import { GridThemeProvider } from "styled-bootstrap-grid"
 import styled from "styled-components"
 import { GlobalStyles } from "Styleguide/Elements/GlobalStyles"
@@ -11,6 +10,7 @@ import { Grid } from "Styleguide/Elements/Grid"
 import { BreakpointVisualizer } from "Styleguide/Utils/BreakpointVisualizer"
 import { Provider as StateProvider } from "unstated"
 import Events from "Utils/Events"
+import { MatchingMediaQueries } from "Utils/Responsive"
 import { ResponsiveProvider } from "Utils/Responsive"
 import { Responsive } from "Utils/Responsive"
