+++ src/Utils/getUser.ts
@@ -1,7 +1,7 @@
-export function getUser(user: User | null): User {
+export function getUser(user: User): User {
   let currentUser = user
 
-  if (currentUser == null) {
+  if (currentUser === undefined) {
     const id = process.env.USER_ID
     const accessToken = process.env.USER_ACCESS_TOKEN
     const labFeatures = process.env.USER_LAB_FEATURES
